Extract annotation parser classes from ReflectionParametricScalar
Parsing annotations is not strictly part of SqlScalar specialization
behavior. This is why it was moved to external class, so it can be reused
in other contexts if required.

ReflectionParametricScalar was renamed to ParametricScalar as it is
no longer bounded to its origin.
